requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
boto3>=1.26.0
python-dotenv==1.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
Flask
Flask-Login
Flask-SQLAlchemy
Werkzeug
